Add files via upload
MuCloud is an application that lets you store your game saves to a cloud storage directory. Please note that you need to have a cloud server installed as a virtual drive to use this app. This feature ensures that your progress is backed up and easily accessible from any location. With MuCloud, you can have peace of mind that your game progress is safe and secure.